backend/requirements.txt
Django==1.11.26
django-extensions==2.2.5
djangorestframework==3.9.4
pytz==2019.3
six==1.13.0
typing==3.7.4.1
